# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user/createUser:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: passWord
        type: string
      - description: 确认密码
        in: formData
        name: repassWord
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: passWord
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 登录
      tags:
      - 用户模块
  /user/getUserList:
    post:
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 所有用户
      tags:
      - 用户模块
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"